requirements.txt
discord.py==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
aiohttp==3.9.1
lxml==4.9.3
openai==1.3.7
